Allow VolumeMounts to be specified in Container.
This extends the knative/serving API to allow for ConfigMaps and Secrets to be mounted as volumes.  All other forms of volume are disallowed.

The container's volume mounts must include all of the provided volumes, they must be ReadOnly (this is defaulted in the webhook), and only a MountPath may be specified.

These are then passed through to the Deployment, and the rest is up to K8s.

Related: https://github.com/knative/serving/issues/1504